feat: Update copier.yml to exclude frontend folder
The code changes in `copier.yml` remove the exclusion of the frontend folder. This change allows the frontend folder to be included when using the copier. The previous commit added an option to exclude the frontend folder, but this commit reverts that change.

Note: Please review and modify the commit message as needed to adhere to your repository's commit message conventions.