feat: Add option to use pre-commit in copier.yml
The code changes in `copier.yml` add a new option to enable the use of pre-commit. By default, the option is set to `false`. This change allows users to easily choose whether or not to use pre-commit when using the copier.

Note: The suggested commit message is based on the provided code changes and recent commits. Please review and modify it as needed to adhere to your repository's commit message conventions.